Refactoring

+++ internal/dto/notification.go
@@ -1,7 +1,7 @@
 package dto
 
 type AddThingNotificationRequest struct {
-	ThingID          int    `json:"thing_id" validate:"required"`
+	ThingID          uint64 `json:"thing_id" validate:"required"`
 	NotificationDate string `json:"notification_date" validate:"required" format:"date-time"`
 }
 
@@ -10,7 +10,7 @@ type UpdateThingNotificationRequest struct {
 }
 
 type ThingNotificationResponse struct {
-	ThingID          int    `json:"id"`
+	ThingID          uint64 `json:"id"`
 	NotificationDate string `json:"notification_date"`
 	CreatedAt        string `json:"created_at"`
 	UpdatedAt        string `json:"updated_at"`
@@ -21,8 +21,8 @@ type ThingNotificationsExtResponse struct {
 }
 
 type ThingNotificationExtResponse struct {
-	ThingID          int    `json:"thing_id"`
-	PlaceID          int    `json:"place_id"`
+	ThingID          uint64 `json:"thing_id"`
+	PlaceID          uint64 `json:"place_id"`
 	ThingTitle       string `json:"thing_title"`
 	PlaceTitle       string `json:"place_title"`
 	NotificationDate string `json:"notification_date"`
