+++ internal/mappers/place_image.go
@@ -2,7 +2,7 @@ package mappers
 
 import "git.dmitriygnatenko.ru/dima/homethings/internal/models"
 
-func ToAddPlaceImageRequest(placeID int, image string) models.AddPlaceImageRequest {
+func ToAddPlaceImageRequest(placeID uint64, image string) models.AddPlaceImageRequest {
 	return models.AddPlaceImageRequest{
 		PlaceID: placeID,
 		Image:   image,
