+++ internal/models/thing_notification.go
@@ -3,25 +3,25 @@ package models
 import "time"
 
 type ThingNotification struct {
-	ThingID          int
+	ThingID          uint64
 	NotificationDate time.Time
 	CreatedAt        time.Time
 	UpdatedAt        time.Time
 }
 
 type AddThingNotificationRequest struct {
-	ThingID          int
+	ThingID          uint64
 	NotificationDate time.Time
 }
 
 type UpdateThingNotificationRequest struct {
-	ThingID          int
+	ThingID          uint64
 	NotificationDate time.Time
 }
 
 type ExtThingNotification struct {
-	ThingID          int
-	PlaceID          int
+	ThingID          uint64
+	PlaceID          uint64
 	ThingTitle       string
 	PlaceTitle       string
 	NotificationDate time.Time
